commiting changes for photo

+++ EssenceRealty.Domain/ViewModel/PhotoViewModel.cs
@@ -13,7 +13,7 @@ namespace EssenceRealty.Domain.ViewModels
         public string UserFilename { get; set; }
         public string Type { get; set; }
         public int? Height { get; set; }
-        public long? Filesize { get; set; }
+        public int? Filesize { get; set; }
         public string Url { get; set; }
         public string Thumb1024 { get; set; }
         public string Thumb180 { get; set; }
